requirements.txt
Flask==1.1.2
line_bot_sdk==2.1.0
python-dotenv==0.19.2
requests==2.12.4
gunicorn
